# server/public/openApi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/cms/category/list": {
      "get": {
        "operationId": "CategoryController_findList",
        "summary": "获取分类列表",
        "parameters": [
          {
            "name": "pageNum",
            "required": false,
            "in": "query",
            "description": "页码",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "每页条数",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "categoryName",
            "required": false,
            "in": "query",
            "description": "分类名称",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "状态：0禁用 1启用",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResultData"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/CategoryEntity"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "CMS-分类管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/cms/category/all": {
      "get": {
        "operationId": "CategoryController_findAll",
        "summary": "获取所有启用分类",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResultData"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/CategoryEntity"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "CMS-分类管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/cms/category/{id}": {
      "get": {
        "operationId": "CategoryController_findOne",
        "summary": "获取分类详情",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResultData"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/CategoryEntity"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "CMS-分类管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "CategoryController_delete",
        "summary": "删除分类",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResultData"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "value": {
                              "type": "boolean",
                              "default": true
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "CMS-分类管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/cms/category": {
      "post": {
        "operationId": "CategoryController_create",
        "summary": "新增分类",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResultData"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "value": {
                              "type": "boolean",
                              "default": true
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "CMS-分类管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "CategoryController_update",
        "summary": "修改分类",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResultData"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "value": {
                              "type": "boolean",
                              "default": true
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "CMS-分类管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/cms/page/list": {
      "get": {
        "operationId": "PageController_findList",
        "summary": "获取单页列表",
        "parameters": [
          {
            "name": "pageNum",
            "required": false,
            "in": "query",
            "description": "页码",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "每页条数",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "pageName",
            "required": false,
            "in": "query",
            "description": "页面名称",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageKey",
            "required": false,
            "in": "query",
            "description": "页面标识",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResultData"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "list": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/PageEntity"
                              }
                            },
                            "total": {
                              "type": "number",
                              "default": 0
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "CMS-单页管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/cms/page/{id}": {
      "get": {
        "operationId": "PageController_findOne",
        "summary": "获取单页详情",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResultData"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PageEntity"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "CMS-单页管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "PageController_delete",
        "summary": "删除单页",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResultData"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "value": {
                              "type": "boolean",
                              "default": true
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "CMS-单页管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/cms/page": {
      "post": {
        "operationId": "PageController_create",
        "summary": "新增单页",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResultData"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "value": {
                              "type": "boolean",
                              "default": true
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "CMS-单页管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "PageController_update",
        "summary": "修改单页",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResultData"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "value": {
                              "type": "boolean",
                              "default": true
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "CMS-单页管理"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Nest-Admin",
    "description": "Nest-Admin 接口文档",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "components": {
    "securitySchemes": {
      "Authorization": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header",
        "name": "Authorization",
        "description": "请在请求头中携带 JWT 令牌，格式：Bearer <token>"
      }
    },
    "schemas": {
...
      "ResultData": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "default": 200
          },
          "msg": {
            "type": "string",
            "default": "操作成功"
          }
        },
        "required": [
          "code",
          "msg"
        ]
      },
...
      "CategoryEntity": {
        "type": "object",
        "properties": {}
      },
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "categoryName": {
            "type": "string",
            "description": "分类名称"
          },
          "categoryCode": {
            "type": "string",
            "description": "分类编码"
          },
          "parentId": {
            "type": "number",
            "description": "父分类ID",
            "default": 0
          },
          "sortOrder": {
            "type": "number",
            "description": "排序",
            "default": 0
          },
          "status": {
            "type": "string",
            "description": "状态：0禁用 1启用",
            "default": "1"
          },
          "icon": {
            "type": "string",
            "description": "图标"
          }
        },
        "required": [
          "categoryName",
          "categoryCode"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "categoryName": {
            "type": "string",
            "description": "分类名称"
          },
          "categoryCode": {
            "type": "string",
            "description": "分类编码"
          },
          "parentId": {
            "type": "number",
            "description": "父分类ID",
            "default": 0
          },
          "sortOrder": {
            "type": "number",
            "description": "排序",
            "default": 0
          },
          "status": {
            "type": "string",
            "description": "状态：0禁用 1启用",
            "default": "1"
          },
          "icon": {
            "type": "string",
            "description": "图标"
          },
          "categoryId": {
            "type": "number",
            "description": "分类ID"
          }
        },
        "required": [
          "categoryName",
          "categoryCode",
          "categoryId"
        ]
      },
...
      "PageEntity": {
        "type": "object",
        "properties": {}
      },
      "CreatePageDto": {
        "type": "object",
        "properties": {
          "pageName": {
            "type": "string",
            "description": "页面名称"
          },
          "pageKey": {
            "type": "string",
            "description": "页面标识"
          },
          "content": {
            "type": "string",
            "description": "页面内容"
          }
        },
        "required": [
          "pageName",
          "pageKey"
        ]
      },
      "UpdatePageDto": {
        "type": "object",
        "properties": {
          "pageName": {
            "type": "string",
            "description": "页面名称"
          },
          "pageKey": {
            "type": "string",
            "description": "页面标识"
          },
          "content": {
            "type": "string",
            "description": "页面内容"
          },
          "pageId": {
            "type": "number",
            "description": "页面ID"
          }
        },
        "required": [
          "pageName",
          "pageKey",
          "pageId"
        ]
      },
...
    }
  }
}